R3 AR-1: move _definitions guard SQL into catalog::writes
Relocate the pure-SQL existence/collision guard builders
(definitions_table_guard_select, existence_guard_select,
rename_collision_guard_select) out of parse::native_sql and into a new
catalog::writes submodule, next to the table identity (DEFINITIONS_TABLE)
and canonical "does not exist" wording (view_not_found_msg) they mirror.
This puts the SQL that guards writes to _definitions with the catalog
module that owns the table, and leaves the parse layer to consume it.

- src/catalog.rs becomes src/catalog/mod.rs; new src/catalog/writes.rs
  holds the three guard builders and their wording unit tests (moved from
  the parse test module).
- parse::native_sql imports the guards from crate::catalog::writes under
  its existing `#[cfg(feature = "extension")]` gate; the DEFINITIONS_TABLE
  import narrows accordingly (the always-compiled guards that referenced
  the schema/table-name consts moved out).

No behaviour change. Verified: the full `just test-all` gate — cargo test
(1045), just test-sql (64/64), and just test-ducklake-ci (6/6) — plus
just build (extension), clippy -D warnings, and fmt all pass.
